Conflicts:
	Makefile.flags

+++ coreutils/du.c
@@ -26,11 +26,7 @@
 //usage:#define du_trivial_usage
 //usage:       "[-aHLdclsx" IF_FEATURE_HUMAN_READABLE("hm") "k] [FILE]..."
 //usage:#define du_full_usage "\n\n"
-//usage:       "Summarize disk space used for each FILE and/or directory.\n"
-//usage:       "Disk space is printed in units of "
-//usage:	IF_FEATURE_DU_DEFAULT_BLOCKSIZE_1K("1024")
-//usage:	IF_NOT_FEATURE_DU_DEFAULT_BLOCKSIZE_1K("512")
-//usage:       " bytes.\n"
+//usage:       "Summarize disk space used for each FILE and/or directory\n"
 //usage:     "\n	-a	Show file sizes too"
 //usage:     "\n	-L	Follow all symlinks"
 //usage:     "\n	-H	Follow symlinks on command line"
@@ -40,11 +36,13 @@
 //usage:     "\n	-s	Display only a total for each argument"
 //usage:     "\n	-x	Skip directories on different filesystems"
 //usage:	IF_FEATURE_HUMAN_READABLE(
-//usage:     "\n	-h	Sizes in human readable format (e.g., 1K 243M 2G )"
+//usage:     "\n	-h	Sizes in human readable format (e.g., 1K 243M 2G)"
 //usage:     "\n	-m	Sizes in megabytes"
 //usage:	)
-//usage:     "\n	-k	Sizes in kilobytes"
-//usage:			IF_FEATURE_DU_DEFAULT_BLOCKSIZE_1K(" (default)")
+//usage:     "\n	-k	Sizes in kilobytes" IF_FEATURE_DU_DEFAULT_BLOCKSIZE_1K(" (default)")
+//usage:	IF_NOT_FEATURE_DU_DEFAULT_BLOCKSIZE_1K(
+//usage:     "\n		Default unit is 512 bytes"
+//usage:	)
 //usage:
 //usage:#define du_example_usage
 //usage:       "$ du\n"
@@ -91,7 +89,7 @@ struct globals {
 #define INIT_G() do { } while (0)
 
 
-static void print(unsigned long size, const char *filename)
+static void print(unsigned long long size, const char *filename)
 {
 	/* TODO - May not want to defer error checking here. */
 #if ENABLE_FEATURE_HUMAN_READABLE
@@ -105,15 +103,15 @@ static void print(unsigned long size, const char *filename)
 		size++;
 		size >>= 1;
 	}
-	printf("%lu\t%s\n", size, filename);
+	printf("%llu\t%s\n", size, filename);
 #endif
 }
 
 /* tiny recursive du */
-static unsigned long du(const char *filename)
+static unsigned long long du(const char *filename)
 {
 	struct stat statbuf;
-	unsigned long sum;
+	unsigned long long sum;
 
 	if (lstat(filename, &statbuf) != 0) {
 		bb_simple_perror_msg(filename);
@@ -190,7 +188,7 @@ static unsigned long du(const char *filename)
 int du_main(int argc, char **argv) MAIN_EXTERNALLY_VISIBLE;
 int du_main(int argc UNUSED_PARAM, char **argv)
 {
-	unsigned long total;
+	unsigned long long total;
 	int slink_depth_save;
 	unsigned opt;
